requirements.txt
flask
google-generativeai
gspread
oauth2client
gunicorn